LARgenerators.bool3d — Function

bool3d(assembly)

Arguments

assembly::LinearAlgebraicRepresentation.Struct

Returns

W::Matrix{Float64}

copEV::SparseMatrixCSC{Int8, Int64}

copFE::SparseMatrixCSC{Int8, Int64}

copCF::SparseMatrixCSC{Int8, Int64}

boolmatrix::BitMatrix

Usage

using LARgenerators, SparseArrays, Base.Threads
import LinearAlgebraicRepresentation as Lar

n, m, p = 1, 1, 1
V, (VV, EV, FV, CV) = Lar.cuboidGrid([n, m, p], true)
cube = V, FV, EV

assembly = Lar.Struct([ cube,
    Lar.t(.3,.4,.25), Lar.r(pi/5,0,0), Lar.r(0,0,pi/12), cube,
    Lar.t(-.2,.4,-.2), Lar.r(0,pi/5,0), Lar.r(0,pi/12,0), cube
])

W, copEV, copFE, copCF, boolmatrix = LARgenerators.bool3d(assembly)

Explanation

After the arrangement, extract all the d-cells from (d-1)-coboundary as isolated polyhedra.

Then compute a single interior point for each of them.

Then compare each such point against all input boundaries, in order to compute those which it was interior to. Extend this point membership as 3-cell containment within the relative input solids.

The point membership with a boundary consists in the parity count of the intersection points of a vertical ray starting at the test point, with the boundary surface.

LARgenerators.internalpoints — Function

internalpoints(V, copEV, copFE, copCF)

Arguments

V::Matrix{Float64}

copEV::SparseMatrixCSC{Int8, Int64}

copFE::SparseMatrixCSC{Int8, Int64}

copCF::SparseMatrixCSC{Int8, Int64}

Return

innerpoints::Vector{Any}

Explanation

Transform each 3-cell in a solid

Compute, for each 3-cell, one internal point

LARgenerators.chainbasis2solids — Function

chainbasis2solids(copEV, copFE, copCF)

Arguments

copEV::SparseMatrixCSC{Int8, Int64}

copFE::SparseMatrixCSC{Int8, Int64}

copCF::SparseMatrixCSC{Int8, Int64}

Return

(FVs, CF)::Tuple{Vector{Vector{Vector{Int64}}}, Vector{Vector{Int64}}}

LARgenerators.getinternalpoint — Function

getinternalpoint(V,EV,FV,Fs, copEV,copFE,z,ptest1,ptest2)

Arguments

V::Matrix{Float64}

EV::Vector{Vector{Int64}}

FV::Vector{Vector{Int64}}

Fs::Vector{Int64}

copEV::SparseMatrixCSC{Int8, Int64}

copFE::SparseMatrixCSC{Int8, Int64}

z::Vector{Any}

ptest1::Vector{Int64}

ptest2::Vector{Int64}

Return

ptest1::Vector{Int64} or ptest2::Vector{Int64}

Explanation

For each test point compute the face planes intersected by vertical ray

Transform each plane in 2D and look whether the intersection point is internal

Return the test point with odd numeber of ray intersections

LARgenerators.testinternalpoint — Function

testinternalpoint(V, EV, FV)

Arguments

V::Matrix{Float64}

EV::Vector{Vector{Int64}}

FV::Vector{Vector{Int64}}

Return

intersectedfaces::Vector{Int64}

LARgenerators.settestpoints — Function

settestpoints(V, FV, Fs, copEV, copFE)

Arguments

V::Matrix{Float64}

FV::Vector{Vector{Int64}}

Fs::Vector{Int64}

copEV::SparseMatrixCSC{Int8, Int64}

copFE::SparseMatrixCSC{Int8, Int64}

Return

(ptest1, ptest2)::Tuple{Vector{Float64}, Vector{Float64}}

LARgenerators.planemap — Function

planemap(V,copEV,copFE,face)(point)

Arguments

V::Matrix{Float64}

copEV::SparseMatrixCSC{Int8, Int64}

copFE::SparseMatrixCSC{Int8, Int64}

face::Int64

Explanation

Tranform the 3D face and the 3D point in their homologous 2D, in order to test for containment.

LARgenerators.rayintersection — Function

rayintersection(point3d::Array{Float64})(V,FV,face::Int)

Explanation

Compute the intersection point of the vertical line through point3d w face. If the face is parallel to z axis return false.

Usage

julia> V,(VV,EV,FV,CV) = Lar.simplex(3,true);

julia> V
3×4 Array{Float64,2}:
 0.0  1.0  0.0  0.0
 0.0  0.0  1.0  0.0
 0.0  0.0  0.0  1.0

julia> FV
4-element Array{Array{Int64,1},1}:
 [1, 2, 3]
 [1, 2, 4]
 [1, 3, 4]
 [2, 3, 4]

 julia> Lar.rayintersection([.333,.333,0])(V,FV,4)
 3-element Array{Float64,1}:
  0.333
  0.333
  0.3340000000000001

LARgenerators.spaceindex — Function

spaceindex(point3d)(model)

Compute the set of face boxes of possible intersection with a point-ray. Work in 3D, where the ray direction is parallel to the z-axis. Return an array of indices of face.

Usage

julia> V,(VV,EV,FV,CV) = Lar.cuboidGrid([1,1,1],true)

julia> spaceindex([.5,.5,.5])((V,FV))
3-element Array{Int64,1}:
 5
 6