LARgenerators.bool2d — Function

bool2d(assembly)

Arguments

assembly::LinearAlgebraicRepresentation.Struct

Return

W::Matrix{Float64}

copEV::SparseMatrixCSC{Int8,Int64}

copFE::SparseMatrixCSC{Int8,Int64}

boolmatrix::BitMatrix

Usage

using SparseArrays, LARgenerators, Base.Threads, SparseArrays, IntervalTrees, LinearAlgebra
import LinearAlgebraicRepresentation as Lar

V, (VV, EV, FV) = Lar.cuboidGrid([1, 1], true)
square = V, EV

assembly = Lar.Struct([
    Lar.Struct([Lar.t(0,0), Lar.r(0), square])
    Lar.Struct([Lar.t(0,0.1), Lar.r(0.1), square])
    Lar.Struct([Lar.t(0,0.2), Lar.r(0.2), square])
    Lar.Struct([Lar.t(0,0.3), Lar.r(0.3), square])
    Lar.Struct([Lar.t(0,0.4), Lar.r(0.4), square])
])

W, copEV, copFE, boolmatrix = LARgenerators.bool2d(assembly)

LARgenerators.testinternalpoint2d — Function

testinternalpoint2d(listOfModels)(testpoint)

Given a point, it returns the list of models containing it.

Arguments

listOfModels::Vector{Any}

testpoint::Vector{Float64}

Return

intersectedFaces: Vector{Int64}

LARgenerators.internalpoints2d — Function

internalpoints2d(W, copEV, copFE)

Returns, for each atom of the decomposition, an internal point.

Arguments

W::Matrix{Float64}

copEV::SparseMatrixCSC{Int8, Int64}

copFE::SparseMatrixCSC{Int8, Int64}

Return

internalpoints::Vector{Vector{Float64}}

LARgenerators.getinternalpoint2d — Function

getinternalpoint2d(W, f, copEV, copFE)

Returns an interior point in the polygon.

Arguments

W::Matrix{Float64}

f::Int64

copEV::SparseMatrixCSC{Int8, Int64}

copFE::SparseMatrixCSC{Int8, Int64}

Return

ptest1::Vector{Float64} or ptest2::Vector{Float64}

LARgenerators.settestpoints2d — Function

settestpoints2d(W, f, copEV, copFE)

Given a polygon, it returns two points, one inside and one outside it.

Arguments

W::Matrix{Float64}

f::Int64

copEV::SparseMatrixCSC{Int8, Int64}

copFE::SparseMatrixCSC{Int8, Int64}

Return

(ptest1, ptest2)::Tuple{Vector{Float64}, Vector{Float64}}

LARgenerators.pointInPolygonClassificationBool — Function

pointInPolygonClassificationBool(V, EV)(pnt)

Determines whether a point is inside or outside a polygon. It is a specific version used for the bool2d function.

Arguments

V::Adjoint{Float64, Matrix{Float64}}

EV::Vector{Vector{Int64}}

pnt::Vector{Float64}

Return

p_in::String or p_out::String

LARgenerators.setTileBool — Function

setTileBool(box)(point)

Returns the tile code of an edge.

Arguments

box::Tuple{Float64, Float64, Float64, Float64}

point::Vector{Float64}

Return

tileCodeBool::Int64